refactor(OtpRequestProcessor): Move class to controllers.api package.
