+++ src/main/java/org/opentripplanner/middleware/OtpMiddlewareMain.java
@@ -13,7 +13,7 @@ import org.opentripplanner.middleware.controllers.api.LogController;
 import org.opentripplanner.middleware.controllers.api.MonitoredTripController;
 import org.opentripplanner.middleware.controllers.api.TripHistoryController;
 import org.opentripplanner.middleware.docs.PublicApiDocGenerator;
-import org.opentripplanner.middleware.otp.OtpRequestProcessor;
+import org.opentripplanner.middleware.controllers.api.OtpRequestProcessor;
 import org.opentripplanner.middleware.persistence.Persistence;
 import org.opentripplanner.middleware.utils.ConfigUtils;
 import org.slf4j.Logger;
